+++ CHANGELOG.md
@@ -19,6 +19,8 @@
   via the `ignore_parent_selectors` configuration option
 * Fix `PropertyUnits` for properties that have multiple unit-like values (e.g.
   shorthand properties) and quoted values
+* Add `SpaceAfterVariableName` linter which checks that there are no spaces
+  between a variable name and a colon
 
 ## 0.39.0